// src/js/model/modifiers.ts
export interface Modifiers {
    noFail: boolean;
    oneLife: boolean;
    fourLives: boolean;
    noBombs: boolean;
    noWalls: boolean;
    noArrows: boolean;
    ghostNotes: boolean;
    disappearingArrows: boolean;
    smallNotes: boolean;
    proMode: boolean;
    strictAngles: boolean;
    zenMode: boolean;
    slowerSong: boolean;
    fasterSong: boolean;
    superFastSong: boolean;
}